msm: camera: isp: Set vfe clock rate
Set the initial vfe clock rate to SVS during probe. When the
clock is enabled the clock driver expects a clock rate to be
voted. The clock is enabled during device open and the
user driver can send the actual rate only after open. This causes
a warning print from clock driver when the clock is enabled in
device open. To avoid the error print set an initail rate.

CRs-Fixed: 1000316
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Change-Id: I66466c2d93b65be07970739eeed6b8156cb7277b
Signed-off-by: Lau <laststandrighthere@gmail.com>