msm: camera: cpp: Vote AHB with clock enabled and no mem retention
Vote for AHB bandwidth after clocks are enabled. For bandwidth
voting clocks need to be on. Set the memory retention flags for
certain hardware clocks to save power, memory does not need to
be retained for these clocks since the driver reinitializes the
jpeg hardware whenever the clocks are turned on.

CRs-Fixed: 1044777 1044373
Signed-off-by: Rajakumar Govindaram <rajakuma@codeaurora.org>
Change-Id: Ieb2f0a1f38757b57acdd2c8c8b34b4e6615db45e
Signed-off-by: Gautham Mayyuri <gmayyuri@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>