msm: vidc_3x: increase timeout value for hw response and power collapse
When running concurrent video session, there are some possibilities
that CPU may take more than a second  to schedule hfi work handler.
In those cases, one of the threads which is waiting for the response
is timing-out.

Increase hw response timeout and power collapse timeout will give
more time for hfi work handler to be scheduled and process the response
messages.

CRs-Fixed: 1086284
Signed-off-by: Karthikeyan Periasamy <kperiasa@codeaurora.org>
Change-Id: I54a37f7a1f4dbe7fd681ac6f76a2e04e1d41f479