crypto: drbg - advance output buffer pointer
The CTR DRBG segments the number of random bytes to be generated into
128 byte blocks. The current code misses the advancement of the output
buffer pointer when the requestor asks for more than 128 bytes of data.
In this case, the next 128 byte block of random numbers is copied to
the beginning of the output buffer again. This implies that only the
first 128 bytes of the output buffer would ever be filled.

The patch adds the advancement of the buffer pointer to fill the entire
buffer.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: mcdachpappe <noreference@web.de>
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>
(cherry picked from commit c1296f88a2511be7fb04cbb22955425c351eab53)
Signed-off-by: TogoFire <togofire@mailfence.com>