crypto: lrw - Convert to skcipher
This patch converts lrw over to the skcipher interface.  It also
optimises the implementation to be based on ECB instead of the
underlying cipher.  For compatibility the existing naming scheme
of lrw(aes) is maintained as opposed to the more obvious one of
lrw(ecb(aes)).

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: mcdachpappe <noreference@web.de>
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>
(cherry picked from commit 5451d86a5ded826f12f91d487f3b514d0f790714)
Signed-off-by: TogoFire <togofire@mailfence.com>