crypto: xts - Convert to skcipher
This patch converts xts over to the skcipher interface.  It also
optimises the implementation to be based on ECB instead of the
underlying cipher.  For compatibility the existing naming scheme
of xts(aes) is maintained as opposed to the more obvious one of
xts(ecb(aes)).

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: mcdachpappe <noreference@web.de>
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>
(cherry picked from commit 82f0bc764d583a74e3924751b732d0ee9a9248f6)
Signed-off-by: TogoFire <togofire@mailfence.com>