crypto: api - Do not clear type bits in crypto_larval_lookup
Currently all bits not set in mask are cleared in crypto_larval_lookup.
This is unnecessary as wherever the type bits are used it is always
masked anyway.

This patch removes the clearing so that we may use bits set in the
type but not in the mask for special purposes, e.g., picking up
internal algorithms.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: mcdachpappe <noreference@web.de>
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>
(cherry picked from commit 744a93024d70a9ea4630eece507dcc2709278499)
Signed-off-by: TogoFire <togofire@mailfence.com>