crypto: simd - Add simd skcipher helper
This patch adds the simd skcipher helper which is meant to be
a replacement for ablk helper.  It replaces the underlying blkcipher
interface with skcipher, and also presents the top-level algorithm
as an skcipher.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: mcdachpappe <noreference@web.de>
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>
(cherry picked from commit 49d64c5822b3c73c9a3a152092f6bdfc4b2cd69d)
Signed-off-by: TogoFire <togofire@mailfence.com>