iio: imu: inv_mpu6050: test whoami first and against all known values
SPI bus is never generating error during transfer, so to check if
a chip is correctly connected on a SPI bus we enforce whoami check
to be correct. In this way we can assure SPI probe is failing if
there is no chip connected.

Note that this is really papering over boards that claim the device
is there when it isn't.  Not a bad thing to cope with, but not
necessarily stable material.

Fixes: cec0154556f8 ("iio: inv_mpu6050: Check WHO_AM_I register on probe")
Signed-off-by: Jean-Baptiste Maneyrol <jmaneyrol@invensense.com>
Change-Id: Id2431b3cec1df77dcf5041ba3755e712c1fa5aad
Signed-off-by: Jonathan Cameron <jic23@kernel.org>
(cherry picked from commit a151aedddf12eac4284f60bda96906d9992e730d)
Signed-off-by: TogoFire <togofire@mailfence.com>