gpu: msm_adreno: Compile only Adreno codes specific to chipset
Idea based on YaroST12/Z2_Kernel@5d969c6, but instead of
completely removing unused parts we only compile Adreno codes that
are specific to target qcom chipset(s) we're compiling against.

To be warned, this is only tested against msm8953 and msm8996
devices only, and may be revised when this change is later known to
break other chipset(s).

Signed-off-by: Albert I <krascgq@outlook.co.id>
Change-Id: I5ca2afff9f3c40904580f177ef5943ec69ca01ad
Signed-off-by: Thagoo <lohitgowda56@gmail.com>
Signed-off-by: TogoFire <italomellopereira@gmail.com