cpufreq: schedutil: clear cached_raw_freq when invalidated
The cpufreq_schedutil governor keeps a cache of the last
raw frequency that was mapped to a supported device frequency.
If the next request for a frequency matches the cached
value, the policy's next_freq value is reused. But there
are paths that can update the raw cached value without
updating the next_freq value, and there are paths that
can set the next_freq value without setting the raw
cached value. On those paths, the cached value
must be reset.

The case that has been observed is when a frequency request
reaches sugov_update_commit but is then rejected by to
the sugov_up_down_rate_limit check.

Bug: 109757301
Signed-off-by: John Dias <joaodias@google.com>
Change-Id: If5fd1028705de4e3f4da3665be77185ea822805b
Signed-off-by: Lau <laststandrighthere@gmail.com>