msm: kgsl: Implement a list of imported memory
With ION now using the dma attachment's priv pointer, we need another
way to associate the imported buffers with the mem entries in kgsl.
Create a list of all dma buffers imported into kgsl, identified by
the first page in the scatterlist. Use this list to determine how
many times a given buffer was mapped as an egl surface and as an egl
image.

Signed-off-by: Harshdeep Dhatt <hdhatt@codeaurora.org>
Change-Id: I2e00411ec4402a4978877acf80d3daf664db7619
Signed-off-by: Lynus Vaz <lvaz@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>