cpu_input_boost: Add early wake boost on power button press
Boost the CPU preemptively upon a power button press to improve wake-up
speeds. The power button being pressed does not always translate to a
wake, but the power cost from a such a brief boost should not be a
deal-breaker.

In testing, the device woke up 10 ms faster on average after this
change.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: Ia64a1413af3651e5862b2cb71f75d02e9c2446a2
Signed-off-by: Lau <laststandrighthere@gmail.com>