BACKPORT: kbuild: replace AS=clang with LLVM_IAS=1
commit 7e20e47c70f810d678d02941fa3c671209c4ca97 upstream.

The 'AS' variable is unused for building the kernel. Only the remaining
usage is to turn on the integrated assembler. A boolean flag is a better
fit for this purpose.

AS=clang was added for experts. So, I replaced it with LLVM_IAS=1,
breaking the backward compatibility.

 /* The commit "kbuild: use LLVM integrated assembler for C files" has been applied. So don't make the mistake of doing this:
ifneq ($(LLVM_IAS),1)
KBUILD_AFLAGS	+= $(CLANG_FLAGS) -no-integrated-as
endif

Doing so will result in this error:
error: unknown target CPU 'cortex-a53'
*/

Suggested-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
Change-Id: I6b7aaab6c0ebdc06934135e6dbc2b1966f49624a
Reviewed-by: Nathan Chancellor <natechancellor@gmail.com>
Reviewed-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry-picked from commit 7e20e47c70f810d678d02941fa3c671209c4ca97)
Signed-off-by: TogoFire <italomellopereira@gmail.com>