qos: Replace expensive cpumask usage with raw bitwise operations
cpumask_set_cpu() uses the set_bit() helper, which, in typical kernels
prior to 4.19, uses a spin lock to guarantee atomicity. This is
expensive and unneeded, especially since the qos functions are hot code
paths. The rest of the cpumask functions use the bitmap API, which is
also more expensive than just doing some simple operations on a word.

Since we're operating with a CPU count that can fit within a word,
replace the expensive cpumask operations with raw bitwise operations
wherever possible to make the pm_qos framework more efficient.

Change-Id: 430548616a5a0333f5a4f1b70e139f14e92d619f
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: TogoFire <italomellopereira@gmail.com>