sched: Do not use IPIs for remote wakeups if idle
Do not wake up and idle CPU for the chance of a cache hit. This avoids
unnecessary energy consumption. Plus, if we need to wake an idle CPU to
handle a remote wakeup, we will induce additional scheduling latency
since the remote CPU will be unavailable while leaving an idle state.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Change-Id: Id4d2477f30bfe260efb4b85eb45c004e6e6c2975