sched/walt: Avoid taking rq lock for every IRQ update
sched_account_irqtime() is called for every IRQ/SoftIRQ update.
The rq->lock is needed only in idle context in which case
update_task_ravg() is called. For IRQ load updates, rq->lock
is not needed, since these are tracked per-cpu and migrations
are not applicable for irqload. By not taking rq lock for
every interrupt, we don't delay the irq exit path under heavy
rq lock contention. For example, this CPU is loaded with thousands
of tasks and rq->lock is acquired during load balance.

Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Change-Id: I506c91ddd054f5945c79f5f8d32b3343233512d7
Signed-off-by: Lau <laststandrighthere@gmail.com>