mm: Eliminate d_path_outlen() and further speed up show_map_vma()
d_path_outlen() isn't needed because we know that d_path() always
populates the given buffer backwards starting from the last byte; with
this, we can easily calculate the length of the generated string by
using the returned pointer from d_path() and the size of the buffer
given to d_path(). This eliminates the need for d_path_outlen() and
removes the bizarre strlen() usage, which makes things simpler and
faster. We also now avoid a memmove() when d_path() completely uses up
its provided buffer.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: Idc47163a6a6ad9cae31f76cc04981298c1c68124
Signed-off-by: TogoFire <togofire@mailfence.com>