bluebooth: Compile out debug logs when !CONFIG_DEBUG_KERNEL
Each time when phone is connected to bluetooth, a large amount
of unnecessary logs get dumped at the dmesg. Hence, guard the
define declaraion of BTFMSLIM_INFO with an if/else macro check of
CONFIG_DEBUG_KERNEL as the logs are only useful when debugging and
we aren't debugging the kernel right now.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: Ie0ca4506807656fb8f85128ec7b4785eaf337bff
Signed-off-by: TogoFire <togofire@mailfence.com>