techpack: ASoC: Don't check if a static array is NULL
A `static` array (in the `.bss` section of a typical executable) always
has a location in memory. It makes no sense to ever check for a `NULL`
here which triggers `-Werror=address` warning. Since this array holds
pointers, it was perhaps "intended" to check individual fields of this
array before dereferencing, which we now do inside the
`msm_routing_find_topology_on_index()` function.

(Other code in this file sets/unsets `cal_data`, but it's not
zero-initialized though.)

Picked from : https://github.com/sonyxperiadev/kernel-techpack-audio/pull/45
Signed-off-by: Marijn Suijten <marijn.suijten@somainline.org>
Change-Id: I492c996d9aee501d15c29a2c86821047ee728be9
Signed-off-by: kawaaii <kawaaii@nocturn9x.space>
[ Tashar02: Modify commit message and description ]
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: TogoFire <togofire@mailfence.com>