defconfig: Disable clock driver for APM X-Gene SoC
The "CONFIG_COMMON_CLK_XGENE" setting in the kernel refers to common clock support for the X-Gene platform. We have SDM625/SDM636/SDM660 for our devices. Processor is not based on the X-Gene architecture, therefore it does not require this configuration in the kernel. Having the clock driver for
APM XGene enabled on an irrelevant platform is simply useless and
redundant. Disable it.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: Ib4457777fffcaf491007294ae60452452469147b
Signed-off-by: TogoFire <togofire@mailfence.com>