leds:qpnp-flash-v2: Add support for flash realtime control
In the default design, every time we change the brightness
in torch node it gets reflected only when the enable bit
in the switch node is toggled from 0 to 1. Add new DT property
to optionally allow brightness change to reflect in realtime,
for as long as the switch node enable bit remains set.

Signed-off-by: Kavya Nunna <quic_knunna@quicinc.com>
Change-Id: I5c0da5ee2ce93a703c4c2b0801bdf28002dda05e
Signed-off-by: TogoFire <togofire@mailfence.com>