leds:qpnp-flash-v2: Add support for flash realtime control
In the default design, every time we change the brightness
in torch node it gets reflected only when the enable bit
in the switch node is toggled from 0 to 1. Add new DT property
to optionally allow brightness change to reflect in realtime,
for as long as the switch node enable bit remains set.

Signed-off-by: Kavya Nunna <quic_knunna@quicinc.com>
Change-Id: I9ca03b277400c949df00cc98bac4467f0d533dde
Signed-off-by: TogoFire <togofire@mailfence.com>