fingerprint: fpc: Mark all functions as inline
An inline function is one for which the compiler copies the code from the
function definition directly into the code of the calling function rather
than creating a separate set of instructions in memory. This is mostly done
for small computations, means a function having less codes so that we can
ignore its overhead and perform faster execution.

If the function is identified as big or complex, compiler will ignore the
inline instruction. So, just mark all of the functions as inline and provide
hints to the compiler and let the compiler decide what should be optimized.
There's no harm in that.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: I75be2fd9e34230f3887caa5cbb6634f95bc18bfc
Signed-off-by: TogoFire <togofire@mailfence.com>