subsystem_restart: Do not use WQ_CPU_INTENSIVE in an unbound workqueue allocation
As per documentation [1], WQ_CPU_INTENSIVE is useful for bound work items which
are expected to hog CPU cycles so that their execution is regulated by the system
scheduler and thereby do not contribute to the concurrency level which is pointless
for an unbound workqueue.
Therefore, using WQ_CPU_INTENSIVE with WQ_BOUND is redundant. Remove it.

[1]: https://www.kernel.org/doc/html/latest/core-api/workqueue.html#flags

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: Id63a09de4ce6849866d91d6cbe9c43aacbcf7084
Signed-off-by: TogoFire <togofire@mailfence.com>