usb: gadget: configfs: Fix mismerge in string handling
Commit 394ca034bb3eca17c173ccb7b175ab8f685264b8 ("usb: gadget: configfs: Fix KASAN use-after-free") has resolved
a use-after-free issue that made use of strcpy() but accidentally it was never imported,
resulting in breaking the original functionality of the patch.

strlcpy() is not suitable here since it may result in truncated strings, causing potential
issues with memory access and data integrity. Therefore, replace it with the originally
intended strcpy().

Fixes: d73e635a21034ea7449762b7455bc8c746da8f51 ("Merge 4.9.263 into android-4.9-q")
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Change-Id: I99f649b5fcc12230b1fb94a0365d5e308c29aba8
Signed-off-by: TogoFire <togofire@mailfence.com>